travis: removing warnings when 9.6 is used

+++ src/alpha_shape/src/alpha.c
@@ -272,8 +272,8 @@ Datum alphashape(PG_FUNCTION_ARGS) {
   PGR_DBG("Strange stuff doing\n");
   funcctx = SRF_PERCALL_SETUP();
 
-  call_cntr = funcctx->call_cntr;
-  max_calls = funcctx->max_calls;
+  call_cntr = (uint32_t)funcctx->call_cntr;
+  max_calls = (uint32_t)funcctx->max_calls;
   tuple_desc = funcctx->tuple_desc;
   res = (vertex_t*) funcctx->user_fctx;
